Solve publishing recursive joint traj problem
The former code will check repeatedly if it is close to the middle point, which may cause problem that it will back to middle point if it not move fast enough and rush to goal point. 
The updated code solve this problem by add a check node for pass the middle point or not.